uint to uint256

+++ contracts/BridgeEth.sol
@@ -6,7 +6,7 @@ import './BridgeBase.sol';
 contract BridgeEth is BridgeBase {
   constructor(address token) BridgeBase(token) {}
 
-  function tokenTransfer(address to, uint amount, uint256 otherChainNonce) external nonReentrant onlyOwner {
+  function tokenTransfer(address to, uint256 amount, uint256 otherChainNonce) external nonReentrant onlyOwner {
     require(processedNonces[otherChainNonce] == false, 'transfer already processed');
     processedNonces[otherChainNonce] = true;
     token.transfer(to, amount);
